Force a coarse move based on current piezo voltage
The current implementation only checked whether or not the requested position
was within a reasonable range of the piezo center, without considering if the
piezo was currently at an extreme. Now, if the current piezo voltage is not
within cOperatingRegion * cPiezoRange of the center, a coarse move will be
forced.